cri: optimize file container_checkpoint.go structure
rename container_checkpoint.go to container_checkpoint_other.go
container_checkpoint_linux.go to container_checkpoint.go
